Update compliance to new checks #1519
